Fixes for unstable net connectivity.
